# perf(nhcb): scrape optimize