Related to #13529 

In `util/convertnhcb` optimize for the happy case when the exposer lists the buckets in ascending order. Reduce the number of allocation by
- reusing the temporary buckets (added `Reset()` function to TempHistogram),
- directly encode spans and bucket counts into the target histogram.

Other small changes:
- use `Compact(2)` to align with the spans vs buckets optimization in [client_golang](https://github.com/prometheus/client_golang/blob/aef8aedb4b6e1fb8ac1c90790645169125594096/prometheus/histogram.go#L1485),
- do `p.tempLsetNHCB = convertnhcb.GetHistogramMetricBase(lset, suffix)` only once in `NHCBParser.processClassicHistogramSeries` as it yields the same result through the histogram series.