@krajorama 2024-10-28
Benchmarks:

Without keep classic histograms:
```
                                                               │ main_nhcb_no_keep_classic.txt │    pr_nhcb_no_keep_classic.txt     │
                                                               │            sec/op             │   sec/op     vs base               │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                            330.7µ ± 6%   333.7µ ± 3%        ~ (p=0.589 n=6)
Parse/data=promtestdata.txt/parser=xpfmt-2                                         1.699m ± 5%   1.715m ± 7%        ~ (p=0.589 n=6)
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                     251.5µ ± 2%   258.6µ ± 3%   +2.86% (p=0.015 n=6)
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                                  1.183m ± 3%   1.152m ± 5%   -2.66% (p=0.015 n=6)
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                       74.41µ ± 5%   72.53µ ± 4%   -2.53% (p=0.041 n=6)
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                               106.45µ ± 3%   79.16µ ± 2%  -25.64% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                              4.523m ± 1%   3.593m ± 1%  -20.56% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                           158.5µ ± 5%   105.4µ ± 2%  -33.48% (p=0.002 n=6)
geomean                                                                            418.9µ        372.5µ       -11.08%

                                                               │ main_nhcb_no_keep_classic.txt │     pr_nhcb_no_keep_classic.txt     │
                                                               │              B/s              │     B/s       vs base               │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                           96.22Mi ± 6%   95.34Mi ± 3%        ~ (p=0.589 n=6)
Parse/data=promtestdata.txt/parser=xpfmt-2                                        18.73Mi ± 5%   18.55Mi ± 7%        ~ (p=0.589 n=6)
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                    95.83Mi ± 2%   93.16Mi ± 3%   -2.79% (p=0.015 n=6)
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                                 20.37Mi ± 3%   20.92Mi ± 5%   +2.72% (p=0.013 n=6)
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                      49.29Mi ± 5%   50.56Mi ± 4%   +2.57% (p=0.041 n=6)
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                               40.11Mi ± 3%   53.94Mi ± 2%  +34.47% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                             7.038Mi ± 1%   8.855Mi ± 1%  +25.81% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                          25.22Mi ± 4%   37.91Mi ± 2%  +50.34% (p=0.002 n=6)
geomean                                                                           32.47Mi        36.51Mi       +12.45%

                                                               │ main_nhcb_no_keep_classic.txt │     pr_nhcb_no_keep_classic.txt     │
                                                               │             B/op              │     B/op      vs base               │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                           58.91Ki ± 0%   59.05Ki ± 0%   +0.24% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=xpfmt-2                                        536.8Ki ± 0%   536.8Ki ± 0%        ~ (p=0.197 n=6)
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                    58.09Ki ± 0%   58.23Ki ± 0%   +0.24% (p=0.002 n=6)
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                                 381.1Ki ± 0%   381.1Ki ± 0%        ~ (p=0.524 n=6)
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                      30.17Ki ± 0%   30.31Ki ± 0%   +0.47% (p=0.002 n=6)
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                               22.91Ki ± 0%   14.05Ki ± 0%  -38.68% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                             110.9Ki ± 0%   111.0Ki ± 0%   +0.13% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                          42.58Ki ± 0%   22.49Ki ± 0%  -47.19% (p=0.002 n=6)
geomean                                                                           83.15Ki        72.31Ki       -13.03%

                                                               │ main_nhcb_no_keep_classic.txt │     pr_nhcb_no_keep_classic.txt      │
                                                               │           allocs/op           │  allocs/op   vs base                 │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                            1.041k ± 0%   1.041k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=promtestdata.txt/parser=xpfmt-2                                         12.35k ± 0%   12.35k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                      831.0 ± 0%    831.0 ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                                  10.01k ± 0%   10.01k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                        545.0 ± 0%    545.0 ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                                 324.0 ± 0%    237.0 ± 0%  -26.85% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                              2.495k ± 0%   2.495k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                            524.0 ± 0%    355.0 ± 0%  -32.25% (p=0.002 n=6)
geomean                                                                            1.493k        1.368k        -8.40%
¹ all samples are equal
```

With keep classic:
```
                                                               │ main_nhcb_keep_classic.txt │      pr_nhcb_keep_classic.txt      │
                                                               │           sec/op           │   sec/op     vs base               │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                         327.6µ ± 3%   334.2µ ± 3%        ~ (p=0.180 n=6)
Parse/data=promtestdata.txt/parser=xpfmt-2                                      1.695m ± 1%   1.696m ± 2%        ~ (p=0.937 n=6)
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                  252.5µ ± 2%   260.7µ ± 1%   +3.24% (p=0.002 n=6)
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                               1.151m ± 7%   1.159m ± 4%        ~ (p=0.937 n=6)
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                    115.5µ ± 3%   116.5µ ± 5%        ~ (p=0.937 n=6)
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                            104.08µ ± 3%   81.34µ ± 5%  -21.85% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                           4.234m ± 1%   3.775m ± 1%  -10.86% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                        157.9µ ± 5%   108.2µ ± 2%  -31.46% (p=0.002 n=6)
geomean                                                                         435.7µ        400.6µ        -8.06%

                                                               │ main_nhcb_keep_classic.txt │      pr_nhcb_keep_classic.txt       │
                                                               │            B/s             │     B/s       vs base               │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                        97.11Mi ± 3%   95.21Mi ± 3%        ~ (p=0.180 n=6)
Parse/data=promtestdata.txt/parser=xpfmt-2                                     18.77Mi ± 1%   18.76Mi ± 2%        ~ (p=0.905 n=6)
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                 95.41Mi ± 2%   92.42Mi ± 1%   -3.13% (p=0.002 n=6)
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                              20.94Mi ± 7%   20.80Mi ± 4%        ~ (p=0.937 n=6)
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                   31.75Mi ± 3%   31.47Mi ± 5%        ~ (p=0.937 n=6)
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                            41.02Mi ± 3%   52.50Mi ± 5%  +27.97% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                          7.515Mi ± 1%   8.426Mi ± 1%  +12.12% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                       25.31Mi ± 5%   36.93Mi ± 2%  +45.91% (p=0.002 n=6)
geomean                                                                        31.22Mi        33.95Mi        +8.75%

                                                               │ main_nhcb_keep_classic.txt │      pr_nhcb_keep_classic.txt       │
                                                               │            B/op            │     B/op      vs base               │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                        58.91Ki ± 0%   59.05Ki ± 0%   +0.24% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=xpfmt-2                                     536.8Ki ± 0%   536.8Ki ± 0%        ~ (p=0.119 n=6)
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                 58.09Ki ± 0%   58.23Ki ± 0%   +0.24% (p=0.002 n=6)
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                              381.1Ki ± 0%   381.1Ki ± 0%        ~ (p=0.898 n=6)
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                   41.17Ki ± 0%   41.31Ki ± 0%   +0.34% (p=0.002 n=6)
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                            22.91Ki ± 0%   14.05Ki ± 0%  -38.68% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                          110.9Ki ± 0%   111.0Ki ± 0%   +0.13% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                       42.58Ki ± 0%   22.49Ki ± 0%  -47.18% (p=0.002 n=6)
geomean                                                                        86.44Ki        75.17Ki       -13.04%

                                                               │ main_nhcb_keep_classic.txt │       pr_nhcb_keep_classic.txt       │
                                                               │         allocs/op          │  allocs/op   vs base                 │
Parse/data=promtestdata.txt/parser=promtext_with_nhcb-2                         1.041k ± 0%   1.041k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=promtestdata.txt/parser=xpfmt-2                                      12.35k ± 0%   12.35k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=promtestdata.nometa.txt/parser=promtext_with_nhcb-2                   831.0 ± 0%    831.0 ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=promtestdata.nometa.txt/parser=xpfmt-2                               10.01k ± 0%   10.01k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=createTestProtoBuf()/parser=promproto_with_nhcb-2                     875.0 ± 0%    875.0 ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2                              324.0 ± 0%    237.0 ± 0%  -26.85% (p=0.002 n=6)
Parse/data=promtestdata.txt/parser=omtext_with_nhcb-2                           2.495k ± 0%   2.495k ± 0%        ~ (p=1.000 n=6) ¹
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2                         524.0 ± 0%    355.0 ± 0%  -32.25% (p=0.002 n=6)
geomean                                                                         1.584k        1.451k        -8.40%
¹ all samples are equal
```